Move sparse_root driver out of __init__.py
